Fix slice access and assignment for blue lists
Python 3 silently removed the `sq_slice` and `sq_ass_slice` slots. As in, they are now a `void*` that's ignored and no longer a function pointer as expected. The suggested work-around is to implement the `mp_subscript` and `mp_ass_subscript` MappingMethods for sequences, so this is what we're doing here.
Last but not least this also adds a missing test case for the slice assignment.