Implement support for `__dict__`
This is a little ugly, but works for the currently known cases. Essentially, what happened is this: The exposed `TypeObject` contains neither `tp_members` nor `tp_methods` entries. Instead, it implements a custom `tp_getattr` (which is deprecated), which relied on the `__members__` magic method (which was deprecated in Python 2.2).
As of Python 3 then the built-in `dir` function no longer uses `__members__` (it was removed) but queries the instance and class `__dict__` attributes (usually these come from a `TypeObject`'s `tp_dict` attribute). Therefore, it was necessary to add support for `__class__` in our custom `tp_getattr` method, as well as change the list returned by `__members__` to be a dictionary for `__dict__`. However, this is currently a suboptimal approach, and _very likely_ not entirely correct as it mostly just fills in dummy values for the `__dict__` entries. But it makes the tests pass, so there's that.
Alas, the real test of this implementation will happen once more real-world code is thrown at it. But even aside from any issues discovered that way, some effort should be taken to register exposed methods and attribute in the `TypeObject`'s relevant attributes. That would pave a way towards a simpler custom implementation of `tp_getattr` (which in turn should also be replaced with a `tp_getattro` implementation).