Fix `BlueList.count()`
For unknown reasons the existing implementation failed after the Python 3 upgrade, but passed with Python 2. However, judging by the way that this was implemented it should have failed with Python 2 as well. This is because the underlying call to `FindKey` will fail with a `RangeCheck` on the last element, as the passed in `keyFrom` parameter will equal the list's size.

In any case, this commit provides an alternate implementation that doesn't require a call to `FindKey` as I did not want to add another special case scenario to `FindKey`.