Update test discovery to work on macOS
The test module was missing the buildflavor mechanism. Not sure why this didn't cause issues before, but it became apparent when trying to run tests on macOS. Alas, the buildflavor is passed onto the tests via an environment variable because the unittest module cannot accept additional command-line arguments.
The PYTHONPATH had to be extended because the built python executable on macOS doesn't find the standard library otherwise. I guess this works on Windows because we do have a Python 3 installation there, but the macOS build agents are currently missing that. Additionally, the tests would pick up wrong binary extensions on macOS (and probably also Windows, but again, we may just have got lucky there so far with the existing Python installation).

Last, but not least, this requires an updated pythonConfig in the branch that exposes the `PYTHON_EXECUTABLE` variable. That config was submitted into Perforce in CL 2331641.