Keep blue livelist disabled by default
This has not been used in years, and is causing us issues with running python interpreter mode in debug, where it was enabled by default. Essentially, the combination of livelist tracking and static variables causes crashes on shutdown because there is no guaranteed order in which clean up happens. This is also one reason why our engine simply hard terminates its process when exiting in game mode.

In any case, the feature is now disabled by default, because we would like to be able to run interpreter mode even in debug configuration, and because nobody has used the livelist feature regularly in the past decade or so.