Fix dir() results on object instances
The Python27 dir() function is more elaborate than Python3 dir() function. It falls back to a default implementation of __dir__, whereas Python3 expects all objects to implement a "dirfunc" called "__dir__", which is reasonable because in Python3 the base object class has a __dir__ method. In the case of the Destiny ballpark, the __dir__ method on the Python class is called, which results in all the Python defined attributes showing up, but none of the C++ ones.

This seems to be caused by this part of the 2.7 generic dir function code, whose functionality has been deprecated in Python 3. It reads the __members__ and __methods__ attributes on the object and merges them into the the dict used for the __dir__ result. The 2.7 version of our codebase seems to have been relying on this, since now things are not showing up in the dir method call (also, the __members__ and __methods__ attributes were properly populated).

    /* Merge in __members__ and __methods__ (if any).
     * This is removed in Python 3000. */
    if (merge_list_attr(dict, obj, "__members__") < 0)
        goto error;
    if (merge_list_attr(dict, obj, "__methods__") < 0)
        goto error;

The dir method call on the base object seems to access the objects __dict__ attribute. This was causing a number of issues since there are a number of things missing from our __dict__.
- Firstly, we usually early out when getting an attribute that exists on the Python object. This is problematic with the __dict__ attribute since we'll end up only getting __dir__ results for the python component of the object, not for the C++ "klass".
- Secondly, even with pure C++ objects, when we get their __dict__ attributes, we currently only return dict entries for variable attributes, but we completely ignore methods.
- Thirdly, once we fix up the first two issues we still end up missing entries for Python Thunkers (The old deprecated way to make an exposure). There are some methods that are attached to interfaces (i.e. IRoot and IList), which end up missing from the __dict__ unless we add some special handling.