Add blue superclass attributes and methods to dir() results
Since we are using mMethodTable to serve both as the blue method table and as the tp_members field of the typeObject, we need to make sure to set the mPyMethodTable field of _classinfo after calling BlueRegisterMethodDefs. We also don't need to traverse parent classes when getting all available methods for a class anymore.