Fix TypeError when checking if Python object in BlueList
This raising of the TypeError has been around in the codebase for a long time, but it only just started becoming an issue with the Python 3 upgrade. In Python 2.7, the Py_TPFLAGS_HAVE_SEQUENCE_IN flag was missing on the list type objects, causing the sq_contains field to be ignored even if it was set on the PySequenceMethods struct. The flag has since been deprecated and so the sq_contains field started working as originally intended, causing PyseqContains_ to be used for checking if the list contains a certain Python object. Since Python 2.7 has simply been doing iterative search and not complaining about the type of the object passed in, this has become the expected behavior. Thus we don't raise an error when checking if a non-IRoot object is in the list, but simply return 0, indicating that the value is not in the list. This makes sense since there is no way that a non-BlueWrapper object exists in a list of BlueWrappers.