Add nullptr guard to `BlueExtractWString`
A rare crash in Trinity points to an access violation
in this particular function. We don't have a reproduction
for the crash as the callstack is heavily optimized in
our reports.

The report and crash dumps do not indicate which line
of code is at fault. A likely cause for the access violation
is a nullptr supplied to `PyUnicode_Check` as that macro
is not guarded against null pointers. If the exception crops
up again, whether in here or other argument extraction overrides,
then we'll need to revisit this in more detail by adding guards
or determining exactly under what rare circumstance a null pointer
could be routed into these function.

Reference:
https://ccpgames.atlassian.net/browse/PLAT-6855