Guard against null-byte characters in `BlueExtractWString`
`PyUnicode_AsWideCharString` can raise an exception and return null
if the passed in object contains a null byte character and the function
doesn't get passed a string length parameter. This caused an access violation
later when we attempted to free a memory block referenced
by a nullptr.

The extracted string gets truncated at the zero-byte's
position. This behavior is not ideal as it contradicts
how standard python strings behave. However, it is non-trivial
to replicate in BlueExposure as the code generally assumes
wide strings to be zero-terminated.

Reference:
https://ccpgames.atlassian.net/browse/PLAT-6855