Make it easier to browse BlueLists in Python debugger (#13)
* Add a nice repr method for BlueList

This should help reduce pain when debugging, since it should now be easier to tell how many items are in a BlueList, and very small lists can be viewed in the debugger.

* Make it easier to debug BlueLists in Python debuggers

When looking at a BlueList in a Python Debugger, it shows up as a Python Object with no visible attributes. The items contained within are a black box and cannot be expanded, making the programmer result to converting the BlueList to a Python list, which makes it hard to keep track of where in the object tree in the debugger the current list resides.

Since BlueLists contain lists of IRoot pointers which may not all have or require Python wrappers, and the overhead of maintaining such wrappers would be costly, keeping these exposed on the Python side all the time is not an option.

Therefore, add an attribute to BlueList as well as a method to populate the attribute with the current items in the list, creating Python wrappers. This will make the items currently in the list browsable in the object tree in the debugger.