Fix gtest discovery and execution against prebuilt python
BluePyStrTest test fixture has been extended with `SetUp` and
`TearDown` functionality which sets up the python interpreter for
test execution. This allows test discovery to skip python
initialization.

Python is then able to be initialized using an environment property
passed to the respective gtest which sets the configuration prefix.
This allows us to run the test suite irrespective of the working
directory.