Support custom build configurations when installing dependencies
VCPKG does not support custom build configurations. The top level
configuration needs to propagate to carbon component dependencies
for the following reasons:
 - In order to build against correctly named binaries, following the
  monolith as an example. F.ex, we need to link blueexposure_internal
   against CcpCore_internal.
- Help ensure we're building carbon dependencies in a consistent manner.

While a lot of the compilation settings controlled by
`CcpGlobalSettings.cmake` across our projects would be better served
in a toolchain file, we still need to ensure we link to the correctly
named binaries no matter what, which is why the custom build flavor
needs to propagate down the dependency tree.